Spotbugs and checkstyle fixes.
