+++ db/jdbc/src/main/java/io/helidon/db/jdbc/JdbcDbProviderBuilder.java
@@ -23,12 +23,12 @@ import io.helidon.common.GenericType;
 import io.helidon.common.configurable.ThreadPoolSupplier;
 import io.helidon.common.mapper.MapperManager;
 import io.helidon.config.Config;
+import io.helidon.db.Db;
 import io.helidon.db.DbInterceptor;
 import io.helidon.db.DbMapper;
 import io.helidon.db.DbMapperManager;
-import io.helidon.db.DbStatements;
-import io.helidon.db.Db;
 import io.helidon.db.DbStatementType;
+import io.helidon.db.DbStatements;
 import io.helidon.db.common.InterceptorSupport;
 import io.helidon.db.spi.DbMapperProvider;
 import io.helidon.db.spi.DbProviderBuilder;
