+++ db/mongodb/src/main/java/io/helidon/db/mongodb/MongoDbProviderBuilder.java
@@ -20,12 +20,12 @@ import java.util.Optional;
 import io.helidon.common.GenericType;
 import io.helidon.common.mapper.MapperManager;
 import io.helidon.config.Config;
+import io.helidon.db.Db;
 import io.helidon.db.DbInterceptor;
 import io.helidon.db.DbMapper;
 import io.helidon.db.DbMapperManager;
-import io.helidon.db.DbStatements;
-import io.helidon.db.Db;
 import io.helidon.db.DbStatementType;
+import io.helidon.db.DbStatements;
 import io.helidon.db.common.InterceptorSupport;
 import io.helidon.db.spi.DbMapperProvider;
 import io.helidon.db.spi.DbProviderBuilder;
