+++ examples/db/jdbc/src/main/java/io/helidon/examples/db/jdbc/DbResultSupport.java
@@ -40,9 +40,10 @@ import io.helidon.webserver.ServerResponse;
 import io.helidon.webserver.Service;
 
 /**
- * TODO javadoc.
+ * Support to write {@link io.helidon.db.DbRowResult} directly to webserver.
+ * This result support creates an array of json objects and writes them to the response entity.
  */
-public class DbResultSupport implements Service, Handler {
+public final class DbResultSupport implements Service, Handler {
 
     /** Local logger instance. */
     private static final Logger LOG = Logger.getLogger(DbResultSupport.class.getName());
@@ -50,11 +51,15 @@ public class DbResultSupport implements Service, Handler {
     private static final JsonBuilderFactory JSON = Json.createBuilderFactory(Collections.emptyMap());
     private static final JsonWriterFactory WRITER_FACTORY = Json.createWriterFactory(Collections.emptyMap());
 
-    public static DbResultSupport create() {
-        return new DbResultSupport();
+    private DbResultSupport() {
     }
 
-    private DbResultSupport() {
+    /**
+     * Create a new instance to register with a webserver.
+     * @return {@link io.helidon.webserver.WebServer} {@link io.helidon.webserver.Service}
+     */
+    public static DbResultSupport create() {
+        return new DbResultSupport();
     }
 
     @Override
